Enable '-Wfloat-conversion' for clang: C only
* compiler_clang_warnings.make: Enable this warning for C. Incidentally
enable '-Wno-reserved-macro-identifier': the code it complained about
now uses a workaround.
* fdlibm_expm1.c: Fix the only such conversion in '*.c'.

+++ fdlibm_expm1.c
@@ -203,7 +203,7 @@ double fdlibm_expm1(double x)
             else
                 {hi = x + ln2_hi; lo = -ln2_lo;  k = -1;}
         } else {
-        k  = invln2*x+((xsb==0)?0.5:-0.5);
+        k  = (int32_t)(invln2*x+((xsb==0)?0.5:-0.5));
         t  = k;
         hi = x - t*ln2_hi;      // t*ln2_hi is exact here
         lo = t*ln2_lo;
